requirements.txt
crewai 
langchain 
langchain-openai
openai 
beautifulsoup4 
requests 
textblob 
matplotlib 
wordcloud
python-dotenv
pandas
langdetect
streamlit

